requirements.txt
streamlit
langchain
langchain-community
langchain-groq
langchain-huggingface
sentence-transformers
chromadb
pypdf
python-docx
unstructured
python-dotenv
pandas
plotly
faiss-cpu
